move tsv generation into the Nallo API
